refactor(core): Move getCurrentInjector/setCurrentInjector to primitives package.
This change casts the injector back and forth since all instances of
injector currently don't implement the `retrieve` method. Note that
the retrieve method is seen as optional, so that Angular can revert back to
inject if necessary.
