+++ goldens/public-api/core/primitives/di/index.api.md
@@ -1,0 +1,41 @@
+## API Report File for "@angular/core_primitives_di"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public
+export function getCurrentInjector(): Injector | undefined | null;
+
+// @public
+export interface InjectionToken<T> extends Type<T> {
+    // (undocumented)
+    ɵprov: ɵɵInjectableDeclaration<T>;
+}
+
+// @public (undocumented)
+export interface Injector {
+    // (undocumented)
+    retrieve?<T>(token: InjectionToken<T>, options?: unknown): T | NotFound;
+}
+
+// @public
+export function isNotFound(e: unknown): e is NotFound;
+
+// @public
+export const NOT_FOUND: unique symbol;
+
+// @public
+export type NotFound = typeof NOT_FOUND | NotFoundError;
+
+// @public
+export class NotFoundError extends Error {
+    constructor(message: string);
+}
+
+// @public (undocumented)
+export function setCurrentInjector(injector: Injector | null | undefined): Injector | undefined | null;
+
+// (No @packageDocumentation comment for this package)
+
+```
