+++ packages/core/src/di/injector_compatibility.ts
@@ -24,9 +24,18 @@ import {
 } from './interface/injector';
 import {ProviderToken} from './provider_token';
 import type {HostAttributeToken} from './host_attribute_token';
+import * as di from '@angular/core/primitives/di';
 
 const _THROW_IF_NOT_FOUND = {};
 export const THROW_IF_NOT_FOUND = _THROW_IF_NOT_FOUND;
+
+export function getCurrentInjector(): Injector {
+  return di.getCurrentInjector() as unknown as Injector;
+}
+
+export function setCurrentInjector(injector: Injector | null | undefined): Injector {
+  return di.setCurrentInjector(injector as di.Injector) as unknown as Injector;
+}
 
 /*
  * Name of a property (that we patch onto DI decorator), which is used as an annotation of which
@@ -41,42 +50,22 @@ const NEW_LINE = /\n/gm;
 const NO_NEW_LINE = 'ɵ';
 export const SOURCE = '__source';
 
-/**
- * Current injector value used by `inject`.
- * - `undefined`: it is an error to call `inject`
- * - `null`: `inject` can be called but there is no injector (limp-mode).
- * - Injector instance: Use the injector for resolution.
- */
-let _currentInjector: Injector | undefined | null = undefined;
-
-export function getCurrentInjector(): Injector | undefined | null {
-  return _currentInjector;
-}
-
-export function setCurrentInjector(
-  injector: Injector | null | undefined,
-): Injector | undefined | null {
-  const former = _currentInjector;
-  _currentInjector = injector;
-  return former;
-}
-
 export function injectInjectorOnly<T>(token: ProviderToken<T>): T;
 export function injectInjectorOnly<T>(token: ProviderToken<T>, flags?: InjectFlags): T | null;
 export function injectInjectorOnly<T>(
   token: ProviderToken<T>,
   flags = InjectFlags.Default,
 ): T | null {
-  if (_currentInjector === undefined) {
+  if (getCurrentInjector() === undefined) {
     throw new RuntimeError(
       RuntimeErrorCode.MISSING_INJECTION_CONTEXT,
       ngDevMode &&
         `inject() must be called from an injection context such as a constructor, a factory function, a field initializer, or a function used with \`runInInjectionContext\`.`,
     );
-  } else if (_currentInjector === null) {
+  } else if (getCurrentInjector() === null) {
     return injectRootLimpMode(token, undefined, flags);
   } else {
-    const value = _currentInjector.get(
+    const value = getCurrentInjector().get(
       token,
       flags & InjectFlags.Optional ? null : undefined,
       flags,
