+++ packages/core/test/bundling/animations/bundle.golden_symbols.json
@@ -113,7 +113,7 @@
   "NG_MOD_DEF",
   "NG_PIPE_DEF",
   "NG_PROV_DEF",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -319,6 +319,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -480,7 +481,7 @@
   "setAllInputsForProperty",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setDirectiveInputsWhichShadowsStyling",
