+++ packages/core/test/bundling/cyclic_import/bundle.golden_symbols.json
@@ -82,7 +82,7 @@
   "NG_MOD_DEF",
   "NG_PIPE_DEF",
   "NG_PROV_DEF",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -247,6 +247,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -388,7 +389,7 @@
   "setAllInputsForProperty",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setDirectiveInputsWhichShadowsStyling",
