+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -111,7 +111,7 @@
   "NG_TEMPLATE_SELECTOR",
   "NG_TEMP_TOKEN_PATH",
   "NG_TOKEN_PATH",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -299,6 +299,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -460,6 +461,7 @@
   "init_di3",
   "init_di4",
   "init_di5",
+  "init_di6",
   "init_di_attr",
   "init_di_setup",
   "init_directive",
@@ -548,6 +550,7 @@
   "init_injector",
   "init_injector2",
   "init_injector3",
+  "init_injector4",
   "init_injector_compatibility",
   "init_injector_discovery_utils",
   "init_injector_profiler",
@@ -620,6 +623,7 @@
   "init_node_selector_matcher",
   "init_noop",
   "init_noop2",
+  "init_not_found",
   "init_null_injector",
   "init_observable",
   "init_operators",
@@ -854,7 +858,7 @@
   "setAllInputsForProperty",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setDirectiveInputsWhichShadowsStyling",
