+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -118,7 +118,7 @@
   "NG_PROV_DEF",
   "NG_VALIDATORS",
   "NG_VALUE_ACCESSOR",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -361,6 +361,7 @@
   "getConstant",
   "getControlAsyncValidators",
   "getControlValidators",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -581,7 +582,7 @@
   "setAllInputsForProperty",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setDirectiveInputsWhichShadowsStyling",
