+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -110,7 +110,7 @@
   "NG_PROV_DEF",
   "NG_VALIDATORS",
   "NG_VALUE_ACCESSOR",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -348,6 +348,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -574,7 +575,7 @@
   "setAllInputsForProperty",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setDirectiveInputsWhichShadowsStyling",
