+++ packages/core/test/bundling/hello_world/bundle.golden_symbols.json
@@ -56,7 +56,7 @@
   "NG_INJ_DEF",
   "NG_MOD_DEF",
   "NG_PROV_DEF",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -193,6 +193,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDeclarationTNode",
@@ -312,7 +313,7 @@
   "setActiveConsumer",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setIncludeViewProviders",
   "setInjectImplementation",
