+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -88,7 +88,7 @@
   "NG_INJ_DEF",
   "NG_PIPE_DEF",
   "NG_PROV_DEF",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -254,6 +254,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -418,7 +419,7 @@
   "setActiveConsumer",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setIncludeViewProviders",
