+++ packages/core/test/bundling/injection/bundle.golden_symbols.json
@@ -43,6 +43,7 @@
   "forEachSingleProvider",
   "forwardRef",
   "getClosureSafeProperty",
+  "getCurrentInjector2",
   "getFactoryDef",
   "getInjectableDef",
   "getInjectorDef",
@@ -62,7 +63,7 @@
   "processInjectorTypesWithProviders",
   "resolveForwardRef",
   "setActiveConsumer",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setInjectImplementation",
   "stringify",
   "walkProviderTree",
