+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -76,7 +76,7 @@
   "NG_INJ_DEF",
   "NG_PIPE_DEF",
   "NG_PROV_DEF",
-  "NOT_FOUND",
+  "NOT_FOUND2",
   "NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR",
   "NOT_YET",
   "NO_CHANGE",
@@ -221,6 +221,7 @@
   "getComponentDef",
   "getComponentLViewByIndex",
   "getConstant",
+  "getCurrentInjector2",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
@@ -347,7 +348,7 @@
   "setActiveConsumer",
   "setBindingRootForHostBindings",
   "setCurrentDirectiveIndex",
-  "setCurrentInjector",
+  "setCurrentInjector2",
   "setCurrentQueryIndex",
   "setCurrentTNode",
   "setIncludeViewProviders",
